Initial commit
This is a test page meant to act as a benchmark for color font support
in web browsers. Also intend to compile the results in major browsers
and major operating systems.

Includes several free-to-share fonts of three explicit-color formats.

Compatible with even more fonts;
Bring your own fonts for a more thorough test.